Major reorganization.
This detatches the actual AJAX Chat instance from the mod.  It must now
be manually managed and installed by the admin.  For now, the location
of AJAX Chat is hard-coded as '$boarddir/chat', i.e.
'$boarddir/chat/lib/classes.php' must exist.

Also removed is the ability to have a Shoutbox that is distinct from the
main chat system; the Shoutbox now always acts as a window into the
"real" chat system.  Put another way, "miniChat" mode is now the only mode.

The option to open the shout box into a pop-up is also gone.